Language.Haskell.HsColour

Description

This is a library which colourises Haskell code. It currently has six output formats:

ANSI terminal codes

LaTeX macros

HTML 3.2 with font tags

HTML 4.01 with external CSS.

XHTML 1.0 with internal CSS.

mIRC chat client colour codes.

Documentation

data Output

The supported output formats.

Constructors

TTY
ANSI terminal codes. Equivalent to TTYg Ansi16Colour but left in for backwards compatibility.

TTYg TerminalType
Terminal codes appropriate for the TerminalType.

LaTeX
TeX macros

HTML
HTML with font tags

CSS
HTML with CSS.

ACSS
HTML with CSS and mouseover types.

ICSS
HTML with inline CSS.

MIRC
mIRC chat clients

data ColourPrefs

Colour preferences.

Constructors

ColourPrefs

Fields

keyword, keyglyph, layout, comment, conid, varid, conop, varop, string, char, number, cpp, selection, variantselection, definition :: [Highlight]

hscolour

Arguments

:: Output
Output format.

-> ColourPrefs
Colour preferences (for formats that support them).

-> Bool
Whether to include anchors.

-> Bool
Whether output document is partial or complete.

-> String
Title for output.

-> Bool
Whether input document is literate haskell or not

-> String
Haskell source code.

-> String
Coloured Haskell source code.

Colourise Haskell source code with the given output format.