Language.Haskell.HsColour.Anchors

Documentation

insertAnchors :: [(TokenType, String)] -> [Either Anchor (TokenType, String)]

insertAnchors places an anchor marker in the token stream before the first defining occurrence of any identifier. Here, before means immediately preceding its type signature, or preceding a (haddock) comment that comes immediately before the type signature, or failing either of those, before the first equation.