Language.Haskell.HsColour.ColourHighlight

Documentation

data Colour

Colours supported by ANSI codes.

Constructors

Black

Red

Green

Yellow

Blue

Magenta

Cyan

White

Rgb Word8 Word8 Word8

data Highlight

Types of highlighting supported by ANSI codes (and some extra styles).

Constructors

Normal

Bold

Dim

Underscore

Blink

ReverseVideo

Concealed

Foreground Colour

Background Colour

Italic

base256 :: Integral int => (Word8 -> Word8 -> Word8 -> r) -> int -> r

Convert an integer in the range [0,2^24-1] to its base 256-triplet, passing the result to the given continuation (avoid unnecessary tupleism).

unbase :: Integral int => int -> Word8 -> Word8 -> Word8 -> int

Convert a three-digit numeral in the given (as arg 1) base to its integer value.

rgb24bit_to_xterm256 :: Integral t => Word8 -> Word8 -> Word8 -> t

Approximate a 24-bit Rgb colour with a colour in the xterm256 6x6x6 colour cube, returning its index.

projectToBasicColour8 :: Colour -> Colour

Ap"proxi"mate a 24-bit Rgb colour with an ANSI8 colour. Will leave other colours unchanged and will never return an Rgb constructor value.

hlProjectToBasicColour8 :: Highlight -> Highlight

Lift projectToBasicColour8 to Highlights