Language.Haskell.HsColour.Colourise

Documentation

module Language.Haskell.HsColour.ColourHighlight

data ColourPrefs

Colour preferences.

Constructors

ColourPrefs

Fields

keyword :: [Highlight]

keyglyph :: [Highlight]

layout :: [Highlight]

comment :: [Highlight]

conid :: [Highlight]

varid :: [Highlight]

conop :: [Highlight]

varop :: [Highlight]

string :: [Highlight]

char :: [Highlight]

number :: [Highlight]

cpp :: [Highlight]

selection :: [Highlight]

variantselection :: [Highlight]

definition :: [Highlight]

readColourPrefs :: IO ColourPrefs

Read colour preferences from .hscolour file in the current directory, or failing that, from $HOME/.hscolour, and failing that, returns a default set of prefs.

defaultColourPrefs :: ColourPrefs

colourise :: ColourPrefs -> TokenType -> [Highlight]

Convert token classification to colour highlights.